First ten split out
